SparklerConfig: Upgrade to snakeyaml in progress
