+++ pom.xml
@@ -40,7 +40,7 @@
         <slf4j.version>1.7.12</slf4j.version>
         <osgi.version>6.0.0</osgi.version>
         <felix.version>5.4.0</felix.version>
-        <snakeyaml.version>1.17</snakeyaml.version>
+        <snakeyaml.version>1.18</snakeyaml.version>
         <json.simple.version>1.1.1</json.simple.version>
         <commons.validator.version>1.5.1</commons.validator.version>
         <jbrowserdriver.version>0.16.4</jbrowserdriver.version>
@@ -57,9 +57,10 @@
         <tika.version>1.14</tika.version>
     </properties>
     <modules>
+        <module>sparkler-plugins</module>
         <module>sparkler-api</module>
         <module>sparkler-app</module>
-        <module>sparkler-plugins</module>
+
         <module>sparkler-tests-base</module>
     </modules>
