+++ sparkler-api/src/main/java/edu/usc/irds/sparkler/configUtils/PluginsProps.java
@@ -1,36 +1,7 @@
 package edu.usc.irds.sparkler.configUtils;
 
+import java.util.HashMap;
 
-import edu.usc.irds.sparkler.BaseConfig;
+public class PluginsProps extends HashMap<String, Object> {
 
-import java.util.ArrayList;
-
-public class PluginsProps implements BaseConfig {
-    private ArrayList<String> active;
-    private UrlFilterProps urlFilter;
-    private FetcherJBrowser fetcherJbrowser;
-
-    public ArrayList<String> getActive() {
-        return active;
-    }
-
-    public void setActive(ArrayList<String> active) {
-        this.active = active;
-    }
-
-    public UrlFilterProps getUrlFilter() {
-        return urlFilter;
-    }
-
-    public void setUrlFilter(UrlFilterProps urlFilter) {
-        this.urlFilter = urlFilter;
-    }
-
-    public FetcherJBrowser getFetcherJbrowser() {
-        return fetcherJbrowser;
-    }
-
-    public void setFetcherJbrowser(FetcherJBrowser fetcherJbrowser) {
-        this.fetcherJbrowser = fetcherJbrowser;
-    }
 }
