+++ sparkler-plugins/urlfilter-regex/src/main/java/edu/usc/irds/sparkler/configUtils/UrlFilterProps.java
